+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as Exemplo } from './Exemplo';
+export { default as GeradorTabela } from './GeradorTabela';
